refactored to keep running stats; no longer store history of durations
It was cumbersome for the user to have to specify how many trials they wanted
to record, and sometimes they might not even know or be able to determine how
many they need to record. Keeping running stats (they are updated at each
prof_stop_trial) allows us to simplify the code by removing the
deltas/durations array, and remove the step that requires the user to convert
the p_data into a p_stats object because the stats are stored and kept
up-to-date on the p_data object.

I've checked that the n, min, max, avg values are correct. I'm not convinced
that I'm doing the stdev correctly, or maybe floating errors are accumulating.